+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 - Add ``DateTime.add`` and ``DateTime.sub`` methods.
 - Make ``DateTime.shift/add/sub`` use ``dateutil.relativedelta.relativedelta`` under the hood which brings new shiftable arguments: years and months
 - Drop support for milliseconds from ``DateTime.shift/add/sub``. **breaking change**
+- Make ``DateTime.parse/format`` understand a subset of `Unicode date patterns <http://www.unicode.org/reports/tr35/tr35-19.html#Date_Field_Symbol_Table>`_ by translating them to strptime/strftime format directives.
 
 
 v0.1.2 (2016-07-26)